+++ apollo-biz/src/main/java/com/ctrip/framework/apollo/biz/config/BizConfig.java
@@ -38,8 +38,9 @@ public class BizConfig extends RefreshableConfig {
   private static final int DEFAULT_APPNAMESPACE_CACHE_REBUILD_INTERVAL = 60; //60s
   private static final int DEFAULT_GRAY_RELEASE_RULE_SCAN_INTERVAL = 60; //60s
   private static final int DEFAULT_APPNAMESPACE_CACHE_SCAN_INTERVAL = 1; //1s
-  private static final int DEFAULT_ACCESSKEY_CACHE_SCAN_INTERVAL = 1; //1s
-  private static final int DEFAULT_ACCESSKEY_CACHE_REBUILD_INTERVAL = 60; //60s
+  private static final int DEFAULT_ACCESS_KEY_CACHE_SCAN_INTERVAL = 1; //1s
+  private static final int DEFAULT_ACCESS_KEY_CACHE_REBUILD_INTERVAL = 60; //60s
+  private static final int DEFAULT_ACCESS_KEY_AUTH_TIME_DIFF_TOLERANCE = 60; //60s
   private static final int DEFAULT_RELEASE_MESSAGE_CACHE_SCAN_INTERVAL = 1; //1s
   private static final int DEFAULT_RELEASE_MESSAGE_SCAN_INTERVAL_IN_MS = 1000; //1000ms
   private static final int DEFAULT_RELEASE_MESSAGE_NOTIFICATION_BATCH = 100;
@@ -138,8 +139,9 @@ public class BizConfig extends RefreshableConfig {
   }
 
   public int accessKeyCacheScanInterval() {
-    int interval = getIntProperty("apollo.access-key-cache-scan.interval", DEFAULT_ACCESSKEY_CACHE_SCAN_INTERVAL);
-    return checkInt(interval, 1, Integer.MAX_VALUE, DEFAULT_ACCESSKEY_CACHE_SCAN_INTERVAL);
+    int interval = getIntProperty("apollo.access-key-cache-scan.interval",
+        DEFAULT_ACCESS_KEY_CACHE_SCAN_INTERVAL);
+    return checkInt(interval, 1, Integer.MAX_VALUE, DEFAULT_ACCESS_KEY_CACHE_SCAN_INTERVAL);
   }
 
   public TimeUnit accessKeyCacheScanIntervalTimeUnit() {
@@ -147,12 +149,20 @@ public class BizConfig extends RefreshableConfig {
   }
 
   public int accessKeyCacheRebuildInterval() {
-    int interval = getIntProperty("apollo.access-key-cache-rebuild.interval", DEFAULT_ACCESSKEY_CACHE_REBUILD_INTERVAL);
-    return checkInt(interval, 1, Integer.MAX_VALUE, DEFAULT_ACCESSKEY_CACHE_REBUILD_INTERVAL);
+    int interval = getIntProperty("apollo.access-key-cache-rebuild.interval",
+        DEFAULT_ACCESS_KEY_CACHE_REBUILD_INTERVAL);
+    return checkInt(interval, 1, Integer.MAX_VALUE, DEFAULT_ACCESS_KEY_CACHE_REBUILD_INTERVAL);
   }
 
   public TimeUnit accessKeyCacheRebuildIntervalTimeUnit() {
     return TimeUnit.SECONDS;
+  }
+
+  public int accessKeyAuthTimeDiffTolerance() {
+    int authTimeDiffTolerance = getIntProperty("apollo.access-key.auth-time-diff-tolerance",
+        DEFAULT_ACCESS_KEY_AUTH_TIME_DIFF_TOLERANCE);
+    return checkInt(authTimeDiffTolerance, 1, Integer.MAX_VALUE,
+        DEFAULT_ACCESS_KEY_AUTH_TIME_DIFF_TOLERANCE);
   }
 
   public int releaseMessageCacheScanInterval() {