+++ apollo-configservice/src/main/java/com/ctrip/framework/apollo/configservice/filter/ClientAuthenticationFilter.java
@@ -16,6 +16,7 @@
  */
 package com.ctrip.framework.apollo.configservice.filter;
 
+import com.ctrip.framework.apollo.biz.config.BizConfig;
 import com.ctrip.framework.apollo.configservice.util.AccessKeyUtil;
 import com.ctrip.framework.apollo.core.signature.Signature;
 import com.ctrip.framework.apollo.core.utils.StringUtils;
@@ -42,16 +43,16 @@ public class ClientAuthenticationFilter implements Filter {
 
   private static final Logger logger = LoggerFactory.getLogger(ClientAuthenticationFilter.class);
 
-  private static final Long TIMESTAMP_INTERVAL = 60 * 1000L;
-
+  private final BizConfig bizConfig;
   private final AccessKeyUtil accessKeyUtil;
 
-  public ClientAuthenticationFilter(AccessKeyUtil accessKeyUtil) {
+  public ClientAuthenticationFilter(BizConfig bizConfig, AccessKeyUtil accessKeyUtil) {
+    this.bizConfig = bizConfig;
     this.accessKeyUtil = accessKeyUtil;
   }
 
   @Override
-  public void init(FilterConfig filterConfig) throws ServletException {
+  public void init(FilterConfig filterConfig) {
     //nothing
   }
 
@@ -106,7 +107,8 @@ public class ClientAuthenticationFilter implements Filter {
     }
 
     long x = System.currentTimeMillis() - requestTimeMillis;
-    return x >= -TIMESTAMP_INTERVAL && x <= TIMESTAMP_INTERVAL;
+    long authTimeDiffToleranceInMillis = bizConfig.accessKeyAuthTimeDiffTolerance() * 1000L;
+    return Math.abs(x) < authTimeDiffToleranceInMillis;
   }
 
   private boolean checkAuthorization(String authorization, List<String> availableSecrets,