+++ CHANGES.md
@@ -9,6 +9,7 @@ Apollo 1.10.0
 * [Fix issue that the $ symbol is not used when reading shell variables](https://github.com/ctripcorp/apollo/pull/3890)
 * [Bump xstream from 1.4.17 to 1.4.18](https://github.com/apolloconfig/apollo/pull/3916)
 * [switch apollo.config-service log from warning to info level](https://github.com/ctripcorp/apollo/pull/3884)
+* [Make Access Key Timestamp check configurable](https://github.com/ctripcorp/apollo/pull/3908)
 * [remove ctrip profile](https://github.com/ctripcorp/apollo/pull/3920)
 
 ------------------